Revert "updates with functions"
